modify, delete 권한체크

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -75,12 +75,20 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
+		
+		if(session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
 		
 		Article foundArticle = articleService.getArticleById(id);
 		
 		if(foundArticle == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다.", id));
+		}
+		
+		if(foundArticle.getMemberId() != (int) session.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-B", "해당 게시물에 대한 권한이 없습니다.");
 		}
 		
 		articleService.deleteArticle(id);
@@ -90,7 +98,11 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession session, int id, String title, String body) {
+		
+		if(session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
 		
 		Article foundArticle = articleService.getArticleById(id);
 		
@@ -98,8 +110,13 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다.", id));
 		}
 		
-		articleService.modifyArticle(id, title, body);
+		//이 게시물 수정 가능한가?(세션에 저장된 memberId랑 게시글에 저장된 memberId가 같은지 확인하는)
+		ResultData actorCanModifyRd = articleService.actorCanModify((int) session.getAttribute("loginedMemberId"), foundArticle.getMemberId());
 		
-		return ResultData.from("S-1", Util.f("%d번 게시글이 수정되었습니다.", id));
+		if(actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+		
+		return articleService.modifyArticle(id, title, body);
 	}
 }
